requirements.txt
image==1.5.33
requests==2.25.1
pyyaml==5.4.1
flask==1.1.2
waitress==2.0.0
Pillow==8.1.2
